check if already ran before running

+++ scripts/node-clientside-tools.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
-npm install -g grunt-cli bower
-rm -rf /home/vagrant/.npm
-rm -rf /home/vagrant/tmp
+
+if [ ! -f "/usr/bin/bower" ]; then
+  npm install -g grunt-cli bower
+  rm -rf /home/vagrant/.npm
+  rm -rf /home/vagrant/tmp
+else
+  echo "Bower and grunt already installed"
+fi
